refactor(paniers): retours relecture
- routes panier : virer les responses 400/404 (génériques, pas de
  raison métier) + alléger les descriptions
- utils : retirer le sort en JS, l'ordre createdAt ASC est déjà
  garanti par la query Prisma (orderBy dans l'include)
- déplacer indicateurPublicIdSchema + panierPublicIdSchema dans
  mb-shared/publicIds (cohérence avec referentielPublicIdSchema,
  individuPublicIdSchema, widgetPublicIdSchema)
- composants Card : inline des props (convention projet : type alias
  réservé aux props réutilisées)
