+++ apps/mb-api/src/panier/routes.ts
@@ -1,11 +1,10 @@
 import { createRoute, OpenAPIHono, z } from '@hono/zod-openapi'
-import { errorApiModelSchema } from '@pilote/mb-shared/error'
 import {
   listPaniersQuerySchema,
   panierApiModelSchema,
   panierListApiModelSchema,
-  panierPublicIdSchema,
 } from '@pilote/mb-shared/panier'
+import { panierPublicIdSchema } from '@pilote/mb-shared/publicIds'
 
 import { requireAuthentication } from '@/framework/auth/requireAuthentication'
 import { never } from '@/framework/errors/never'
@@ -15,7 +14,6 @@ import { listPaniers } from '@/panier/queries/listPaniers'
 
 const PanierApiModelSchema = panierApiModelSchema.openapi('PanierApiModel')
 const PanierListApiModelSchema = panierListApiModelSchema.openapi('PanierListApiModel')
-const ErrorApiModelSchema = errorApiModelSchema.openapi('ErrorApiModel')
 
 // --- GET /paniers ------------------------------------------------------------
 
@@ -25,17 +23,13 @@ const getPaniersRoute = createRoute({
   tags: ['Panier'],
   summary: "Lister les paniers d'indicateurs",
   description:
-    "Retourne la liste paginée des paniers d'indicateurs. La pagination est cursor-based : passez `cursor` (renvoyé dans la réponse précédente) pour obtenir la page suivante. Chaque item inclut `indicateurIds`, triés par ordre d'insertion dans le panier (createdAt ASC de la jonction).",
+    "Chaque item inclut `indicateurIds`, triés par ordre d'insertion dans le panier (createdAt ASC de la jonction).",
   middleware: [requireAuthentication],
   request: { query: listPaniersQuerySchema },
   responses: {
     200: {
       content: { 'application/json': { schema: PanierListApiModelSchema } },
       description: 'Liste paginée des paniers',
-    },
-    400: {
-      content: { 'application/json': { schema: ErrorApiModelSchema } },
-      description: 'Paramètres de requête invalides',
     },
   },
 })
@@ -52,17 +46,13 @@ const getPanierByIdRoute = createRoute({
   tags: ['Panier'],
   summary: 'Récupérer un panier par identifiant public',
   description:
-    "Retourne un panier identifié par son identifiant public (format `PAN-XXX`). La réponse inclut `indicateurIds` triés par ordre d'insertion (createdAt ASC de la jonction). Renvoie 404 (`ENTITY_NOT_FOUND`) si aucun panier ne correspond.",
+    "La réponse inclut `indicateurIds` triés par ordre d'insertion (createdAt ASC de la jonction).",
   middleware: [requireAuthentication],
   request: { params: detailParamsSchema },
   responses: {
     200: {
       content: { 'application/json': { schema: PanierApiModelSchema } },
       description: 'Panier trouvé',
-    },
-    404: {
-      content: { 'application/json': { schema: ErrorApiModelSchema } },
-      description: 'Panier introuvable',
     },
   },
 })
