+++ apps/mb-api/src/panier/utils.ts
@@ -3,20 +3,16 @@ import { type PanierApiModel } from '@pilote/mb-shared/panier'
 import { type IndicateurModel, type PanierModel } from '@/generated/prisma/models'
 
 export type PanierWithIndicateurs = PanierModel & {
-  indicateurs: Array<{
-    createdAt: Date
-    indicateur: Pick<IndicateurModel, 'publicId'>
-  }>
+  indicateurs: Array<{ indicateur: Pick<IndicateurModel, 'publicId'> }>
 }
 
+// L'ordre des `indicateurs` est garanti par la query Prisma
+// (`orderBy: { createdAt: 'asc' }` dans l'include).
 export const toPanierApiModel = (panier: PanierWithIndicateurs): PanierApiModel => ({
   id: panier.publicId,
   nom: panier.nom,
   description: panier.description,
-  indicateurIds: panier.indicateurs
-    .slice()
-    .sort((a, b) => a.createdAt.getTime() - b.createdAt.getTime())
-    .map((lien) => lien.indicateur.publicId),
+  indicateurIds: panier.indicateurs.map((lien) => lien.indicateur.publicId),
   createdAt: panier.createdAt.toISOString(),
   updatedAt: panier.updatedAt.toISOString(),
 })
